rename use-order-by.ts -> use-metric-order-by.ts
